Refactor useSecuritySteps
