+++ src/routes/safe/components/Apps/hooks/useSecuritySteps.ts
@@ -1,5 +1,7 @@
-import { useState, useEffect, useCallback } from 'react'
+import { useState, useEffect, useCallback, useMemo } from 'react'
+import { useSafeAppUrl } from 'src/logic/hooks/useSafeAppUrl'
 import { loadFromStorage, saveToStorage } from 'src/utils/storage'
+import { useAppList } from './appList/useAppList'
 
 const APPS_SECURITY_STEPS_APPS_REVIEWED = 'APPS_SECURITY_STEPS_APPS_REVIEWED'
 const APPS_SECURITY_STEPS_CUSTOM_APPS_REVIEWED = 'APPS_SECURITY_STEPS_CUSTOM_APPS_REVIEWED'
@@ -7,20 +9,22 @@ const APPS_SECURITY_STEPS_EXTENDED_LIST_REVIEWED = 'APPS_SECURITY_STEPS_EXTENDED
 const APPS_SECURITY_STEPS_LEGAL_CONSENT_REVIEWED = 'APPS_SECURITY_STEPS_LEGAL_CONSENT_REVIEWED'
 
 const useSecuritySteps = (): {
-  appsReviewed: string[]
-  onReviewApp: (appId: string) => void
-  extendedListReviewed: boolean
-  onReviewExtendedList: () => void
-  customAppsReviewed: string[]
-  onReviewCustomApp: (appUrl: string) => void
-  onRemoveCustomApp: (appUrl: string) => void
+  handleConfirm: (shouldHide: boolean) => void
+  showDisclaimer: boolean
   consentReceived: boolean
-  onConsentReceipt: () => void
+  isSafeAppInDefaultList: boolean
+  isFirstTimeAccessingApp: boolean
+  extendedListReviewed: boolean
+  onRemoveCustomApp: (appUrl: string) => void
 } => {
   const [appsReviewed, setAppsReviewed] = useState<string[]>([])
   const [extendedListReviewed, setExtendedListReviewed] = useState(false)
   const [customAppsReviewed, setCustomAppsReviewed] = useState<string[]>([])
   const [consentReceived, setConsentReceived] = useState<boolean>(false)
+  const [isDisclaimerReadingCompleted, setIsDisclaimerReadingCompleted] = useState(false)
+  const { isLoading, appList, getSafeApp } = useAppList()
+  const { getAppUrl } = useSafeAppUrl()
+  const url = getAppUrl()
 
   useEffect(() => {
     setAppsReviewed(loadFromStorage(APPS_SECURITY_STEPS_APPS_REVIEWED) || [])
@@ -28,6 +32,12 @@ const useSecuritySteps = (): {
     setCustomAppsReviewed(loadFromStorage(APPS_SECURITY_STEPS_CUSTOM_APPS_REVIEWED) || [])
     setConsentReceived(loadFromStorage(APPS_SECURITY_STEPS_LEGAL_CONSENT_REVIEWED) || false)
   }, [])
+
+  useEffect(() => {
+    if (!url) {
+      setIsDisclaimerReadingCompleted(false)
+    }
+  }, [url])
 
   const onReviewApp = useCallback(
     (appId: string): void => {
@@ -72,16 +82,63 @@ const useSecuritySteps = (): {
     [customAppsReviewed],
   )
 
+  const isSafeAppInDefaultList = useMemo(() => {
+    if (!url) return false
+
+    const urlInstance = new URL(url)
+    const safeAppUrl = `${urlInstance.hostname}/${urlInstance.pathname}`
+
+    return appList.some((appItem) => {
+      const appItemUrl = new URL(appItem.url)
+      return `${appItemUrl.hostname}/${appItemUrl.pathname}` === safeAppUrl
+    })
+  }, [appList, url])
+
+  const isFirstTimeAccessingApp = useMemo(() => {
+    if (!url) return true
+
+    const safeAppId = getSafeApp(url)?.id
+
+    return safeAppId ? !appsReviewed?.includes(safeAppId) : !customAppsReviewed?.includes(url)
+  }, [appsReviewed, customAppsReviewed, getSafeApp, url])
+
+  const handleConfirm = useCallback(
+    (shouldHide: boolean) => {
+      onConsentReceipt()
+
+      const safeAppId = getSafeApp(url)?.id
+
+      if (safeAppId) {
+        onReviewApp(safeAppId)
+      } else {
+        if (shouldHide) {
+          onReviewCustomApp(url)
+        }
+      }
+
+      if (!extendedListReviewed) {
+        onReviewExtendedList()
+      }
+
+      setIsDisclaimerReadingCompleted(true)
+    },
+    [extendedListReviewed, getSafeApp, onConsentReceipt, onReviewApp, onReviewCustomApp, onReviewExtendedList, url],
+  )
+
+  const showDisclaimer =
+    !isLoading &&
+    (!consentReceived ||
+      (isSafeAppInDefaultList && isFirstTimeAccessingApp) ||
+      (!isSafeAppInDefaultList && isFirstTimeAccessingApp && !isDisclaimerReadingCompleted))
+
   return {
-    appsReviewed,
-    onReviewApp,
-    extendedListReviewed,
-    onReviewExtendedList,
-    customAppsReviewed,
-    onReviewCustomApp,
-    onRemoveCustomApp,
+    handleConfirm,
+    showDisclaimer,
     consentReceived,
-    onConsentReceipt,
+    isSafeAppInDefaultList,
+    isFirstTimeAccessingApp,
+    extendedListReviewed,
+    onRemoveCustomApp,
   }
 }
